Move song tag foundation under tag namespaces
