+++ src/server/packages/Song/Domain/Models/Tag/SongTagFactoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Song\Domain\Models;
+namespace Song\Domain\Models\Tag;
 
 use Support\Domain\ValueObjects\OrderNo;
